client/requirements.txt
click==7.1.2
grpcio==1.35.0
protobuf==3.14.0
